Add `pyarrow` as required dependency as needed for dagster-duckdb
Before adding this dependency, when one attempts to materialize all assets, they are presented with the following error:

```
ImportError: Unable to find a usable engine; tried using: 'pyarrow', 'fastparquet'.
A suitable version of pyarrow or fastparquet is required for parquet support.
Trying to import the above resulted in these errors:
 - Missing optional dependency 'pyarrow'. pyarrow is required for parquet support. Use pip or conda to install pyarrow.
 - Missing optional dependency 'fastparquet'. fastparquet is required for parquet support. Use pip or conda to install fastparquet.
```

This commit introduces the opinionated change of using `pyarrow` over the alternative of `fastparquet`.